Unifies assistant routes; moves model register
Centralizes assistant routing by using enum-based paths for overview and model registration.

Moves the model registration UI into a components directory and updates related imports.

Removes debug installation effects to reduce console noise.

Improves maintainability and routing consistency.
